+++ hexital/indicators/rsi.py
@@ -35,7 +35,7 @@ class RSI(Indicator[float | None]):
         losses = None
 
         if self.prev_exists():
-            change = self.prev_reading(self.source) - self.reading(self.source)
+            change = self.prev_src() - self.src()
 
             change_gain = -1 * change if change < 0 else 0.0
             change_loss = change if change > 0 else 0.0
@@ -49,7 +49,7 @@ class RSI(Indicator[float | None]):
             ) / self.period
         elif self.reading_period(self.period + 1, self.source):
             changes = [
-                self.reading(self.source, i) - self.reading(self.source, i - 1)
+                self.at_src(i) - self.at_src(i - 1)
                 for i in range(index - (self.period - 1), index + 1)
             ]